Move math source file out of entries
