+++ .babelrc
@@ -1,6 +1,7 @@
 {
   "presets": [
     ["env", {
+      "debug": true,
       "modules": false,
       "targets": {
         "browsers": "> 1%",
