+++ app/javascript/math.js
